+++ js/modules/bootstrap-select.js
@@ -11,9 +11,19 @@ import 'bootstrap-select/dist/js/bootstrap-select';
 
 $(function(){
 
-    $('.selectpicker').selectpicker({
-      style: 'btn-default bs-placeholder',
-      size: 'auto',
+    function initBootstrapSelect() {
+        $('.selectpicker').selectpicker({
+          style: 'btn-default bs-placeholder',
+          size: 'auto',
+        });
+    }
+
+    // on page load
+    initBootstrapSelect();
+
+    // custom event
+    $(window).on('initBootstrapSelect softpage:opened form-modal:opened', function() {
+        initBootstrapSelect();
     });
 
 });