+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Each release is divided into the following main categories:
 
 ### IMPORTANT
 
+- Contact Form: New script added that can be included in `app.js` with `import 'allink-core-static/js/modules/contact-form';`
 - Softpage: `.conent-section` withing a tingle modal were globally resetted to `padding:0` and `margin:0`. Double check the spacings when updating, or ask Beat for advice ;)
 
 ### NEW